feat(auth): improve signup flow and test setup
- Fixed incorrect verification code logic in signup controller
- Updated Playwright config to use new test setup file
- Replaced global test setup with user-specific auth setup
- Enhanced signup test with verification code input
- Added placeholder for OTP input in user setup popup
- Updated localization files with new verification code message
