+++ tests/user.auth.setup.ts
@@ -1,0 +1,40 @@
+import { test, expect } from "@playwright/test";
+import { CONFIGS } from "./config";
+
+test("create storage state", async ({ page }) => {
+  const id = CONFIGS.PLAYWRIGHT.ID;
+  const email = `playwright+${id}@ratel.foundation`;
+  const password = "password1234!@#$";
+  const displayName = `Playwright User ${id}`;
+  const userName = `pw-${id}`;
+  console.log(`🆕 Creating new user: ${email} / ${password}`);
+
+  await page.goto(CONFIGS.PLAYWRIGHT.BASE_URL!);
+  await page.getByRole("button", { name: /sign in/i }).click();
+  await page.getByText("Create an account").click();
+
+  await page.getByPlaceholder("Email", { exact: true }).fill(email);
+  await page.getByText("Send").click();
+
+  await page
+    .getByPlaceholder("Verify code in your email.", { exact: true })
+    .fill("000000");
+  await page.getByText("Verify").click();
+
+  await page.getByPlaceholder(/password/i).fill(password);
+  await page.getByPlaceholder(/display name/i).fill(displayName);
+  await page.getByPlaceholder(/user name/i).fill(userName);
+
+  // Accept terms by clicking the label (checkbox is hidden)
+  const tosCheckbox = page.locator('label[for="agree_checkbox"]');
+  await tosCheckbox.click();
+  await page.getByRole("button", { name: /finished sign-up/i }).click();
+  await expect(page.getByText(/start/i)).toBeVisible();
+
+  // Save Playwright storage state for authenticated tests
+  await page.context().storageState({ path: "user.json" });
+
+  console.log("✅ Global authenticated user setup completed");
+  console.log(`📄 Test user saved: ${email}`);
+  console.log(`🔐 Storage state saved to: user.json`);
+});
